Adicionando os arquivos line. Adicionando os casos triviais da seleção de linha no arquivo selection.c

+++ selection.h
@@ -2,9 +2,10 @@
 #define SELECTION_H
 
 #include "point.h"
+#include "line.h"
 #include "data.h"
 
-int checkPoint(float, float, float, float, int);
+int checkPoint(double, double, int, int, int);
 Point *selectPoint(int, int);
 
 #endif
